(refactor) - put multer config in its own file.
